requirements.txt
fastapi
pydantic
pydantic-settings
sqlalchemy
sqlmodel
asyncpg>=0.29.0
redis>=5.0.4
PyJWT
python-dotenv>=1.0.1
qrcode[pil]>=7.4.2
passlib[bcrypt]
uvicorn>=0.23.0,<1.0.0
asyncmy>=0.2.10
loguru